+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Updated
+
+- `externalId` is generated from the API
+
+### Removed
+
+- **[BREAKING]** `generateExternalId` from `form-service`
+
 ## [16.1.0] - 2024-09-30
 
 ### Changed